# lib/events: Be more resilient against dropping events (fixes #3952)